Update sd-card-osx.md
Reorganizeing content so we can better judge what needs to be done. Please note that we have a much more extensive discussion of this already in the handbook. Documentation does not goo into any motivation about why to use NOOBS or not. This is actually discussed on the NOOBS web page